Re-design `CodeMap` to profit from append-only semantics (#1898)
* move code_map.rs -> code_map/mod.rs

* move EngineFunc[Span][Iter] into its own submodule

* move SmallByteSlice into new utils submodule

* add new CodeMap2 implementation

* drop-in replace CodeMap with CodeMap2

* remove unnecessary attribute

* remove unnecessary bound checks

* change panic messages

* change panic message

* add hot and inline annotations

* add hint::cold_path where necessary

* change hint usage

* fix docs

* add CodeView and make CodeMap fully lock-free

* bump MAX_FUNCS limits

* re-sync CodeView after host func calls

This fixes a soundness issue where host function calls could compile new Wasm modules, create new Wasm instances, and changes Wasm tables that would make it possible to call functions without updating the CodeView.

* derived LEN_BUCKET0 from LEN_BUCKET0_LOG2

* improve const docs

* re-design asserts and their docs

* re-format docs

* add CodeView::features accessor

* impl conversions between EngineFunc and u32

* add Engine::resolve_func method

* make [return_]call_internal ops use FuncEntry refs

This avoids the EngineFunc -> &FuncEntry machinery for those ops which are quite costly for call-intense workloads.

* rename [{Un,C}ompiled]FuncEntity -> [{Un,C}ompiled]FuncEntry

* make UncompiledFuncEntry and assoc APIs private

* make FuncEntry::uninit private

* add safety comments (resolve TODOs)

* deduplicate executor code

* rename CodeView::get_ref -> entry

* rename CodeView::get -> get_or_compile

* refactor and clean-up CodeVIew::get_or_compile

* separate hot and cold paths in FuncEntry::get_or_compile

* avoid inline of imported wasm calls

Imported host calls are more important for performance usually.

* fix doc links

+++ crates/ir/src/encode.rs
@@ -171,7 +171,7 @@ impl_encode_using! {
     Address as u64 = Into::into,
     Func as u32 = Into::into,
     FuncType as u32 = Into::into,
-    InternalFunc as u32 = Into::into,
+    InternalFunc as usize = Into::into,
     Global as u32 = Into::into,
     Memory as u16 = Into::into,
     Table as u32 = Into::into,
